+++ hpc/nvidia/cuda/cuda_adapter.h
@@ -1733,11 +1733,27 @@
  * } __nv_bfloat162_raw;
  */
 #if defined(__USE_HPCC__)
+# define __nv_fp8_e4m3	__hpcc_fp8_e4m3
+# define __nv_fp8_e5m2	__hpcc_fp8_e5m2
+# define __nv_fp8x2_e4m3	__hpcc_fp8x2_e4m3
+# define __nv_fp8x4_e4m3	__hpcc_fp8x4_e4m3
+# define __nv_fp8x2_e5m2	__hpcc_fp8x2_e5m2
+# define __nv_fp8x4_e5m2	__hpcc_fp8x4_e5m2
+# define __nv_fp8x2_e8m0	__hpcc_fp8x2_e8m0
+# define __nv_fp8x4_e8m0	__hpcc_fp8x4_e8m0
 # define __nv_bfloat16	__hpcc_bfloat16
 # define __nv_bfloat16_raw	__hpcc_bfloat16_raw
 # define __nv_bfloat162	__hpcc_bfloat162
 # define __nv_bfloat162_raw	__hpcc_bfloat162_raw
 #elif defined(__USE_HIP__)
+# define __nv_fp8_e4m3	__hip_fp8_e4m3
+# define __nv_fp8_e5m2	__hip_fp8_e5m2
+# define __nv_fp8x2_e4m3	__hip_fp8x2_e4m3
+# define __nv_fp8x4_e4m3	__hip_fp8x4_e4m3
+# define __nv_fp8x2_e5m2	__hip_fp8x2_e5m2
+# define __nv_fp8x4_e5m2	__hip_fp8x4_e5m2
+# define __nv_fp8x2_e8m0	__hip_fp8x2_e8m0
+# define __nv_fp8x4_e8m0	__hip_fp8x4_e8m0
 # define __nv_bfloat16	__hip_bfloat16
 # define __nv_bfloat16_raw	__hip_bfloat16_raw
 # define __nv_bfloat162	__hip_bfloat162